+++ internship/forms/organization_form.py
@@ -23,32 +23,12 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-
-from django import forms
 from django.forms import ModelForm
+
 from internship.models.organization import Organization
-from internship.models.organization_address import OrganizationAddress
-from internship.models.period import Period
-from functools import partial
-DateInput = partial(forms.DateInput, {'class': 'datepicker'})
 
 
 class OrganizationForm(ModelForm):
     class Meta:
         model = Organization
         fields = ['name', 'website', 'reference']
-
-
-class OrganizationAddressForm(ModelForm):
-    class Meta:
-        model = OrganizationAddress
-        fields = ['location', 'postal_code', 'city', 'country', 'latitude', 'longitude']
-
-
-class PeriodForm(ModelForm):
-    class Meta:
-        model = Period
-        fields = ['name', 'date_start', 'date_end']
-        widgets = {'date_start': forms.DateInput(format='%d/%m/%Y'),
-                   'date_end': forms.DateInput(format='%d/%m/%Y'),
-                   }
